pyproject.toml
[project]
name = "vilbot"
version = "0.1.0"
description = "vilbot aecord bot"
authors = [
    {name = "Alex Lipscomb", email = "aecord777@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "py-cord>=2.6.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "pytest>=8.3.2",
    "pre-commit>=3.8.0",
]

[tool.pyright]
include = ["public"]
exclude = ["**/__pycache__", "__pypackages__", ".venv"]
venvPath = "."
venv = ".venv"
executionEnvironments = [{ root = "." }]
